Added first POI and refactored the WeightedList class

+++ Assets/Scripts/Utils/WeightedEntry.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 [Serializable]
 public class WeightedEntry<T> {
     [SerializeField] private T value;
-    [SerializeField] private float weight;
+    [SerializeField] private int weight;
 
-    public WeightedEntry(T value, float weight) {
+    public WeightedEntry(T value, int weight) {
         this.value = value;
         this.weight = weight;
     }
 
     public T Value => value;
 
-    public float Weight => weight;
+    public int Weight => weight;
 }
